+++ src/lib/core/Toaster.svelte
@@ -16,17 +16,18 @@
 	const toasts = $derived(getToasts());
 
 	const positionClasses: Record<ToasterPosition, string> = {
-		'top-left': 'top-4 left-4 items-start',
-		'top-center': 'top-4 left-1/2 -translate-x-1/2 items-center',
-		'top-right': 'top-4 right-4 items-end',
-		'bottom-left': 'bottom-4 left-4 items-start',
-		'bottom-center': 'bottom-4 left-1/2 -translate-x-1/2 items-center',
-		'bottom-right': 'bottom-4 right-4 items-end'
+		'top-left': 'top-4 left-4',
+		'top-center': 'top-4 left-1/2 -translate-x-1/2',
+		'top-right': 'top-4 right-4',
+		'bottom-left': 'bottom-4 left-4',
+		'bottom-center': 'bottom-4 left-1/2 -translate-x-1/2',
+		'bottom-right': 'bottom-4 right-4'
 	};
 </script>
 
 <div
 	class={cn('fixed z-toast flex flex-col gap-2 pointer-events-none', positionClasses[position], className)}
+	style="width: 24rem;"
 >
 	{#each toasts as toast (toast.id)}
 		<Toast {toast} onclose={removeToast} />